requirements.txt
pyzotero==1.5.18
python-slugify==8.0.1
requests==2.32.5
